fix: Update db.py to correct error in result pagination
It seems that the order of arguments sent to the query_text and query_date functions are reversed.  Therefore, on the first search where the intent is for limit to equal 5 and offset to be 0, in reality, offset equals 5 and limit = 0.  With limit = 0, then there is no limit and all results after 5 are returned.  This practically breaks any your agent when there is any significant conversational history because it returns all the messages and will overload your context space.